ready-tensor-course/week3/pyproject.toml
[project]
name = "week3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.29",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
